+++ src/ErrorOr/EmptyErrors.cs
@@ -1,6 +1,0 @@
-namespace ErrorOr;
-
-internal static class EmptyErrors
-{
-    public static List<Error> Instance => new();
-}