+++ src/ErrorOr/ErrorOr.Recordable.cs
@@ -1,31 +1,10 @@
-using System.Text.Json;
-using System.Text.Json.Serialization;
-
-namespace ErrorOr;
+﻿namespace ErrorOr;
 
 public readonly partial record struct ErrorOr<TValue>
 {
     /// <inheritdoc/>
-    public string GetRecording() => GetRecording(RecordableDefaults.JsonOptions);
-
-    /// <inheritdoc/>
-    public string GetRecording(JsonSerializerOptions options)
+    public TOutput GetRecording<TOutput>(IRecordingSerializer<TOutput> serializer)
     {
-        if (IsError)
-        {
-            return JsonSerializer.Serialize(Errors, options);
-        }
-
-        return JsonSerializer.Serialize(Value, options);
+        return IsError ? serializer.SerializeErrors(Errors) : serializer.SerializeValue(Value);
     }
-}
-
-internal static class RecordableDefaults
-{
-    internal static readonly JsonSerializerOptions JsonOptions = new()
-    {
-        WriteIndented = true,
-        Converters = { new JsonStringEnumConverter() },
-        DefaultIgnoreCondition = JsonIgnoreCondition.Never,
-    };
 }