+++ tests/ErrorOr/ErrorOr.ToErrorOrTests.cs
@@ -49,6 +49,21 @@ public class ToErrorOrTests
     }
 
     [Fact]
+    public async Task ErrorToErrorOrAsync_WhenAccessingFirstError_ShouldReturnSameError()
+    {
+        // Arrange
+        Error error = Error.Unauthorized();
+        Task<Error> task = Task.FromResult(error);
+
+        // Act
+        ErrorOr<int> result = await task.ToErrorOrAsync<int>();
+
+        // Assert
+        result.IsError.Should().BeTrue();
+        result.FirstError.Should().Be(error);
+    }
+
+    [Fact]
     public void ListOfErrorsToErrorOr_WhenAccessingErrors_ShouldReturnSameErrors()
     {
         // Arrange
@@ -63,12 +78,60 @@ public class ToErrorOrTests
     }
 
     [Fact]
-    public void ArrayOfErrorsToErrorOr_WhenAccessingErrors_ShouldReturnSimilarErrors()
+    public async Task ListOfErrorsToErrorOrAsync_WhenAccessingErrors_ShouldReturnSameErrors()
     {
+        // Arrange
+        List<Error> errors = [Error.Unauthorized(), Error.Validation()];
+        Task<List<Error>> task = Task.FromResult(errors);
+
+        // Act
+        ErrorOr<int> result = await task.ToErrorOrAsync<int>();
+
+        // Assert
+        result.IsError.Should().BeTrue();
+        result.Errors.Should().BeEquivalentTo(errors);
+    }
+
+    [Fact]
+    public void EnumerableOfErrorsToErrorOr_WhenAccessingErrors_ShouldReturnSimilarErrors()
+    {
+        // Arrange
         Error[] errors = [Error.Unauthorized(), Error.Validation()];
 
+        // Act
         ErrorOr<int> result = errors.ToErrorOr<int>();
 
+        // Assert
+        result.IsError.Should().BeTrue();
+        result.Errors.Should().Equal(errors);
+    }
+
+    [Fact]
+    public async Task ArrayOfErrorsToErrorOrAsync_WhenAccessingErrors_ShouldReturnSimilarErrors()
+    {
+        // Arrange
+        Error[] errors = [Error.Unauthorized(), Error.Validation()];
+        Task<Error[]> task = Task.FromResult(errors);
+
+        // Act
+        ErrorOr<int> result = await task.ToErrorOrAsync<int>();
+
+        // Assert
+        result.IsError.Should().BeTrue();
+        result.Errors.Should().Equal(errors);
+    }
+
+    [Fact]
+    public async Task EnumerableOfErrorsToErrorOrAsync_WhenAccessingErrors_ShouldReturnSimilarErrors()
+    {
+        // Arrange
+        Error[] errors = [Error.Unauthorized(), Error.Validation()];
+        Task<IEnumerable<Error>> task = Task.FromResult((IEnumerable<Error>)errors);
+
+        // Act
+        ErrorOr<int> result = await task.ToErrorOrAsync<int>();
+
+        // Assert
         result.IsError.Should().BeTrue();
         result.Errors.Should().Equal(errors);
     }